backend/requirements.txt
# ==================== 核心框架 ====================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# ==================== 数据库 ====================
psycopg2-binary==2.9.9  # PostgreSQL适配器
mysqlclient==2.2.0  # MySQL适配器（可选）

# ==================== 异步任务 ====================
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ==================== 认证与安全 ====================
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
cryptography==41.0.7

# ==================== AI/LLM相关 ====================
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.2
tiktoken==0.5.2  # Token计数

# ==================== 图像处理 ====================
Pillow==10.1.0
opencv-python==4.8.1.78
imageio==2.33.1
imageio-ffmpeg==0.4.9

# ==================== 视频处理 ====================
moviepy==1.0.3
ffmpeg-python==0.2.0

# ==================== HTTP客户端 ====================
requests==2.31.0
httpx==0.25.2  # 支持异步HTTP
aiohttp==3.9.1  # 异步HTTP客户端

# ==================== 数据处理 ====================
pandas==2.1.4
numpy==1.26.2

# ==================== 文件存储 ====================
boto3==1.34.10  # AWS S3
django-storages==1.14.2

# ==================== 模板引擎 ====================
Jinja2==3.1.2

# ==================== 工具库 ====================
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1

# ==================== 监控与日志 ====================
sentry-sdk==1.39.2  # 错误追踪
python-json-logger==2.0.7

# ==================== 测试 ====================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
factory-boy==3.3.0

# ==================== 代码质量 ====================
black==23.12.1  # 代码格式化
flake8==6.1.0  # 代码检查
isort==5.13.2  # import排序
mypy==1.7.1  # 类型检查

# ==================== 开发工具 ====================
ipython==8.18.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# ==================== ASGI服务器 ====================
uvicorn==0.25.0
gunicorn==21.2.0
daphne==4.0.0

# ==================== 其他依赖 ====================
python-magic==0.4.27  # 文件类型检测
chardet==5.2.0  # 字符编码检测
